separate tag-related functionality into its own module
